# openapi_timesheet/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: fhb190776@fh-vie.ac.at
  description: This is a simple timeshoot tool... (coming soon)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Timesheet - OpenAPI 3.0
  version: 1.0.11
servers:
- url: /
...
paths:
...
  /project/{project-hoursId}:
    delete:
      description: For valid response try integer IDs with value less then 10.
      operationId: deleteproject
      parameters:
      - description: ID of the project-hours that needs to be deleted
        explode: false
        in: path
        name: project-hoursId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: project not found
      summary: Delete project by ID
      tags:
      - project-hours
      x-accepts: application/json
      x-tags:
      - tag: project-hours
    get:
      description: For valid response try integer IDs with value less then 5 or higher
        10.
      operationId: getproject-hoursById
      parameters:
      - description: ID of project that needs to be fetched
        explode: false
        in: path
        name: project-hoursId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
            application/xml:
              schema:
                $ref: '#/components/schemas/Project'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: project-hours not found
      summary: Find Project by ID
      tags:
      - project-hours
      x-accepts: application/json
      x-tags:
      - tag: project-hours
...
components:
...
  schemas:
    Project:
      example:
        name: name
        pid: 10
      properties:
        pid:
          example: 10
          format: int64
          type: integer
        name:
          type: string
      type: object
      xml:
        name: project
...